checkpointing, created model module
